Render pipeline metrics in the new Dashboard
