+++ sematic/ui/packages/common/src/pages/RunDetails/CenterPane.tsx
@@ -3,6 +3,7 @@ import FunctionSection from "src/pages/RunDetails/FunctionSection";
 import RunTabs from "src/pages/RunDetails/RunTabs";
 import { useMemo } from "react";
 import ReactFlowDag from "src/pages/RunDetails/dag/ReactFlowDag";
+import PipelineMetrics from "src/pages/RunDetails/pipelineMetrics";
 
 const CenterPane = () => {
     const { selectedPanel } = useRunDetailsSelectionContext();
@@ -10,6 +11,9 @@ const CenterPane = () => {
     return useMemo(() => {
         if (selectedPanel === "dag") {
             return <ReactFlowDag />;
+        }
+        if (selectedPanel === "metrics") {
+            return <PipelineMetrics />;
         }
         return <>
             <FunctionSection />
